requirements.txt
google-adk>=1.25.0
google-cloud-bigquery>=3.40.0
python-dotenv>=1.0.0
